test(extension): remove tokensPageObject.ts
